+++ Releases.md
@@ -1,5 +1,21 @@
 # Release Updates
 
+## Version 2.3.3
+
+-   Reveal Leaf during initial vault load fix
+
+## Version 2.3.2
+
+-   Style settings plugin implementation
+
+## Version 2.3.1
+
+-   Default Evernote View Change
+
+## Version 2.3.0
+
+-   File sorting options enhancement
+
 ## Version 2.2.9
 
 -   Vertical and Horizontal Evernote Views